requirements.txt
Django==1.11.9
Markdown==2.6.11
Pillow==5.0.0
PyYAML==3.12
certifi==2017.11.5
chardet==3.0.4
click==5.1
compass-stylesheets==0.12.6
django-mustache==1.1.0
djangorestframework==3.7.7
html-json-forms==1.0.0
idna==2.6
natural-keys==1.2.1
pyScss==1.3.5
pystache==0.5.4
pytz==2017.3
requests==2.18.4
six==1.11.0
swapper==1.1.0
urllib3==1.22
wq.app==1.0.0
wq.core==1.0.0
wq.db==1.0.0.post1
